rancher-machine: Fix brew audit RuboCop warning: Style/IfUnlessModifier
Warning was:

    $ brew style lyraphase/right2repair/rancher-machine
    Taps/lyraphase/homebrew-right2repair/Formula/rancher-machine.rb:24:7: C: [Correctable] Style/IfUnlessModifier: Favor modifier unless usage when having a single-line body. Another good alternative is the usage of control flow &&/||.
          unless File.exist?(dest_path)
          ^^^^^^

+++ Formula/rancher-machine.rb
@@ -21,9 +21,7 @@ class RancherMachine < Formula
         s.gsub!("docker_machine", "docker_machine")
       end
       dest_path = path.gsub("docker-machine", "rancher-machine")
-      unless File.exist?(dest_path)
-        File.rename(path, dest_path)
-      end
+      File.rename(path, dest_path) unless File.exist?(dest_path)
     end
 
     bash_completion.install Dir["contrib/completion/bash/*.bash"]
